# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sheriff Department Shift Scheduling API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface AuthUser {
  id: number;
  email: string;
  firstName: string;
  lastName: string;
  role: string;
  /** @nullable */
  shiftId?: number | null;
  /** @nullable */
  shiftName?: string | null;
  isActive: boolean;
}

export interface User {
  id: number;
  email: string;
  firstName: string;
  lastName: string;
  role: string;
  /** @nullable */
  shiftId?: number | null;
  /** @nullable */
  shiftName?: string | null;
  isActive: boolean;
  createdAt: string;
}

export type CreateUserBodyRole =
  (typeof CreateUserBodyRole)[keyof typeof CreateUserBodyRole];

export const CreateUserBodyRole = {
  admin: "admin",
  sergeant: "sergeant",
  deputy: "deputy",
} as const;

export interface CreateUserBody {
  email: string;
  password: string;
  firstName: string;
  lastName: string;
  role: CreateUserBodyRole;
  /** @nullable */
  shiftId?: number | null;
}

export type UpdateUserBodyRole =
  (typeof UpdateUserBodyRole)[keyof typeof UpdateUserBodyRole];

export const UpdateUserBodyRole = {
  admin: "admin",
  sergeant: "sergeant",
  deputy: "deputy",
} as const;

export interface UpdateUserBody {
  email?: string;
  firstName?: string;
  lastName?: string;
  role?: UpdateUserBodyRole;
  /** @nullable */
  shiftId?: number | null;
  isActive?: boolean;
  password?: string;
}

export type ShiftShiftType =
  (typeof ShiftShiftType)[keyof typeof ShiftShiftType];

export const ShiftShiftType = {
  day: "day",
  night: "night",
} as const;

export type ShiftShiftLetter =
  (typeof ShiftShiftLetter)[keyof typeof ShiftShiftLetter];

export const ShiftShiftLetter = {
  a: "a",
  b: "b",
} as const;

export interface Shift {
  id: number;
  name: string;
  shiftType: ShiftShiftType;
  shiftLetter: ShiftShiftLetter;
  /** @nullable */
  sergeantId?: number | null;
  /** @nullable */
  sergeantName?: string | null;
}

export type ShiftWithRosterShiftType =
  (typeof ShiftWithRosterShiftType)[keyof typeof ShiftWithRosterShiftType];

export const ShiftWithRosterShiftType = {
  day: "day",
  night: "night",
} as const;

export type ShiftWithRosterShiftLetter =
  (typeof ShiftWithRosterShiftLetter)[keyof typeof ShiftWithRosterShiftLetter];

export const ShiftWithRosterShiftLetter = {
  a: "a",
  b: "b",
} as const;

export interface ShiftMember {
  id: number;
  userId: number;
  firstName: string;
  lastName: string;
  role: string;
  email: string;
}

export interface ShiftWithRoster {
  id: number;
  name: string;
  shiftType: ShiftWithRosterShiftType;
  shiftLetter: ShiftWithRosterShiftLetter;
  /** @nullable */
  sergeantId?: number | null;
  /** @nullable */
  sergeantName?: string | null;
  members: ShiftMember[];
}

export interface UpdateShiftBody {
  /** @nullable */
  sergeantId?: number | null;
}

export interface ShiftAssignment {
  id: number;
  userId: number;
  shiftId: number;
  firstName: string;
  lastName: string;
  role: string;
  shiftName: string;
  createdAt: string;
}

export interface CreateShiftAssignmentBody {
  userId: number;
  shiftId: number;
}

export type ScheduleDayWorkingShiftLetter =
  (typeof ScheduleDayWorkingShiftLetter)[keyof typeof ScheduleDayWorkingShiftLetter];

export const ScheduleDayWorkingShiftLetter = {
  a: "a",
  b: "b",
} as const;

export interface ScheduleShift {
  id: number;
  name: string;
  shiftType: string;
  shiftLetter: string;
  isWorking: boolean;
  memberCount: number;
  /** @nullable */
  sergeantName?: string | null;
}

export interface ScheduleDay {
  date: string;
  dayOfWeek: string;
  workingShiftLetter: ScheduleDayWorkingShiftLetter;
  shifts: ScheduleShift[];
}

export interface TodaySchedule {
  date: string;
  workingShiftLetter: string;
  dayShift: ShiftWithRoster;
  nightShift: ShiftWithRoster;
}

export type DayOffRequestStatus =
  (typeof DayOffRequestStatus)[keyof typeof DayOffRequestStatus];

export const DayOffRequestStatus = {
  pending: "pending",
  approved: "approved",
  denied: "denied",
} as const;

export interface DayOffRequest {
  id: number;
  userId: number;
  requestedDate: string;
  reason: string;
  status: DayOffRequestStatus;
  /** @nullable */
  reviewedById?: number | null;
  /** @nullable */
  reviewNotes?: string | null;
  /** @nullable */
  reviewedAt?: string | null;
  createdAt: string;
  requesterFirstName: string;
  requesterLastName: string;
  /** @nullable */
  requesterShiftName?: string | null;
  /** @nullable */
  reviewerFirstName?: string | null;
  /** @nullable */
  reviewerLastName?: string | null;
}

export interface CreateDayOffRequestBody {
  requestedDate: string;
  reason: string;
}

export interface ReviewRequestBody {
  reviewNotes?: string;
}

export interface ShiftCount {
  shiftId: number;
  shiftName: string;
  count: number;
}

export interface DashboardSummary {
  totalPersonnel: number;
  activePersonnel: number;
  pendingRequests: number;
  todayWorkingShift: string;
  shiftCounts: ShiftCount[];
  recentRequests: DayOffRequest[];
}

export interface Notification {
  id: number;
  recipientId: number;
  type: string;
  message: string;
  isRead: boolean;
  createdAt: string;
}

export type ListShiftAssignmentsParams = {
  userId?: number;
  shiftId?: number;
};

export type GetScheduleParams = {
  start: string;
  end: string;
};

export type ListDayOffRequestsParams = {
  userId?: number;
  status?: ListDayOffRequestsStatus;
  shiftId?: number;
};

export type ListDayOffRequestsStatus =
  (typeof ListDayOffRequestsStatus)[keyof typeof ListDayOffRequestsStatus];

export const ListDayOffRequestsStatus = {
  pending: "pending",
  approved: "approved",
  denied: "denied",
} as const;