# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sheriff Department Shift Scheduling API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # USERS
...
  /shift-assignments:
    get:
      operationId: listShiftAssignments
      tags: [shiftAssignments]
      summary: List shift assignments
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
        - name: shiftId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Assignments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ShiftAssignment"
    post:
      operationId: createShiftAssignment
      tags: [shiftAssignments]
      summary: Assign a user to a shift
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateShiftAssignmentBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShiftAssignment"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /shift-assignments/{id}:
    delete:
      operationId: deleteShiftAssignment
      tags: [shiftAssignments]
      summary: Remove a shift assignment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # SCHEDULE
...
  /day-off-requests:
    get:
      operationId: listDayOffRequests
      tags: [dayOffRequests]
      summary: List day off requests
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, denied]
        - name: shiftId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Requests list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DayOffRequest"
    post:
      operationId: createDayOffRequest
      tags: [dayOffRequests]
      summary: Submit a day off request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDayOffRequestBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DayOffRequest"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /day-off-requests/{id}:
    get:
      operationId: getDayOffRequest
      tags: [dayOffRequests]
      summary: Get day off request by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DayOffRequest"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /day-off-requests/{id}/approve:
    patch:
      operationId: approveDayOffRequest
      tags: [dayOffRequests]
      summary: Approve a day off request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewRequestBody"
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DayOffRequest"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /day-off-requests/{id}/deny:
    patch:
      operationId: denyDayOffRequest
      tags: [dayOffRequests]
      summary: Deny a day off request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewRequestBody"
      responses:
        "200":
          description: Denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DayOffRequest"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # DASHBOARD
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthUser:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
        shiftId:
          type: ["integer", "null"]
        shiftName:
          type: ["string", "null"]
        isActive:
          type: boolean
      required:
        - id
        - email
        - firstName
        - lastName
        - role
        - isActive
...
    ShiftAssignment:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        shiftId:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
        shiftName:
          type: string
        createdAt:
          type: string
      required:
        - id
        - userId
        - shiftId
        - firstName
        - lastName
        - role
        - shiftName
        - createdAt
    CreateShiftAssignmentBody:
      type: object
      properties:
        userId:
          type: integer
        shiftId:
          type: integer
      required:
        - userId
        - shiftId
...
    DayOffRequest:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        requestedDate:
          type: string
        reason:
          type: string
        status:
          type: string
          enum: [pending, approved, denied]
        reviewedById:
          type: ["integer", "null"]
        reviewNotes:
          type: ["string", "null"]
        reviewedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        requesterFirstName:
          type: string
        requesterLastName:
          type: string
        requesterShiftName:
          type: ["string", "null"]
        reviewerFirstName:
          type: ["string", "null"]
        reviewerLastName:
          type: ["string", "null"]
      required:
        - id
        - userId
        - requestedDate
        - reason
        - status
        - createdAt
        - requesterFirstName
        - requesterLastName
    CreateDayOffRequestBody:
      type: object
      properties:
        requestedDate:
          type: string
        reason:
          type: string
      required:
        - requestedDate
        - reason
    ReviewRequestBody:
      type: object
      properties:
        reviewNotes:
          type: string
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        recipientId:
          type: integer
        type:
          type: string
        message:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - recipientId
        - type
        - message
        - isRead
        - createdAt